Function pointers

+++ 0x0F-function_pointers/1-array_iterator.c
@@ -1,3 +1,4 @@
+#include <stdlib.h>
 #include "function_pointers.h"
 
 /**
@@ -10,12 +11,11 @@
  */
 void array_iterator(int *array, size_t size, void (*action)(int))
 {
-	unsigned int i;
+	size_t i;
 
 	if (action)
 	{
 		for (i = 0; i < size; i++)
 			action(array[i]);
 	}
-		
 }
